requirements.txt
PySide6>=6.4.0
google-generativeai>=0.5.0
httplib2>=0.22.0
certifi>=2023.7.22
python-dotenv>=1.0.0 